+++ src/index.ts
@@ -110,6 +110,9 @@ import { CenterComponent } from './center/center.component';
 
 import { LoadingContainer } from './loading-container/loading-container.component';
 
+import { DateRangeModule } from './date-range/date-range.module';
+import { DateRangeComponent } from './date-range/date-range.component';
+
 import {
   MatToolbarModule,
   MatMenuModule,
@@ -119,6 +122,7 @@ import {
   MatIconModule,
   MatRippleModule,
   MatProgressSpinnerModule,
+  MatNativeDateModule,
 } from '@angular/material';
 
 const components = [
@@ -187,9 +191,11 @@ const components = [
     MatIconModule,
     MatRippleModule,
     MatProgressSpinnerModule,
+    MatNativeDateModule,
+    DateRangeModule,
   ],
   declarations: components,
-  exports: [...components],
+  exports: [...components, DateRangeComponent],
 })
 export class IgniteDesignSystemModule {
   static forRoot(): ModuleWithProviders {